requirements.txt
Flask==3.0.3
gunicorn==23.0.0
WeasyPrint==66.0
Pillow==10.4.0
qrcode[pil]==7.4.2
jinja2==3.1.4
psycopg[binary]==3.2.11
boto3==1.35.0